vm: widen svalue_t::type to uint32_t
The 16 low bits of the svalue type word are fully allocated (T_LVALUE
0x1 through T_LVALUE_CODEPOINT 0x8000), leaving no room for new value
types. Widen the field to 32 bits so new type tags can use 0x10000 and
above; sizeof(svalue_t) is unchanged on 64-bit targets (4 + 2 + 2
padding + 8-byte union).

Also widen the few helpers that carried a runtime type tag in an
unsigned short: the deferred compound-free queue in svalue.cc and the
cycle-walker frames in cycles.cc. Compiler-side variable/function type
words (TYPE_MOD_*/DECL_*) are a separate 16-bit namespace and are
deliberately untouched.

No behavior change; preparation for the T_PROMISE value type (#1319).

+++ src/packages/contrib/cycles.cc
@@ -63,7 +63,7 @@ struct PendingFix {
 };
 
 struct Frame {
-  unsigned short type;  // T_ARRAY, T_CLASS, T_MAPPING, or T_FUNCTION
+  uint32_t type;  // T_ARRAY, T_CLASS, T_MAPPING, or T_FUNCTION
   union {
     array_t* arr;
     mapping_t* map;
@@ -127,7 +127,7 @@ std::string render_key(const svalue_t* key) {
   }
 }
 
-Frame make_frame(unsigned short type, void* ptr) {
+Frame make_frame(uint32_t type, void* ptr) {
   Frame f{};
   f.type = type;
   f.u.ptr = ptr;
@@ -155,7 +155,7 @@ void cycle_walk(svalue_t* root, WalkMode mode, WalkResult* res) {
   // svalue slot; it CAN be a back-edge target, because bind() shares the
   // args array between the old and the new funptr (f_bind in
   // packages/core/efuns_main.cc bumps args->ref instead of copying).
-  auto handle_edge = [&](unsigned short ttype, void* tptr, svalue_t* slot, SlotKind kind,
+  auto handle_edge = [&](uint32_t ttype, void* tptr, svalue_t* slot, SlotKind kind,
                          std::string&& label) {
     auto ins = color.try_emplace(tptr, COLOR_GREY);
     if (ins.second) {  // white: tree edge, descend
